feat: 增加充值功能

+++ src/service/events/bill.ts
@@ -36,6 +36,6 @@ export const pushBill = async ({
       $inc: { balance: -price }
     });
   } catch (error) {
-    Bill.findByIdAndDelete(billId);
+    billId && Bill.findByIdAndDelete(billId);
   }
 };
